228: Extracted switch case for question type enum into new strategy, created corresponding classes for each
