+++ src/main/java/de/imi/mopat/controller/strategy/FreeTextOrBarcodeStrat.java
@@ -1,0 +1,23 @@
+package de.imi.mopat.controller.strategy;
+
+import de.imi.mopat.controller.QuestionController;
+import de.imi.mopat.model.FreetextAnswer;
+import de.imi.mopat.model.Question;
+import de.imi.mopat.model.Questionnaire;
+import de.imi.mopat.model.dto.QuestionDTO;
+import org.springframework.validation.BindingResult;
+
+public class FreeTextOrBarcodeStrat implements CreateOrUpdateAnswerStrategy {
+    @Override
+    public void createOrUpdateAnswer(QuestionDTO questionDTO, Question question, QuestionController controller, BindingResult result, Questionnaire questionnaire) {
+        if (!question.getAnswers().isEmpty()) {
+            // Update freetext answer
+            FreetextAnswer freetextAnswer = (FreetextAnswer) question.getAnswers().get(0);
+            freetextAnswer.setQuestion(question);
+            freetextAnswer.setIsEnabled(true);
+        } else {
+            // Create new answer
+            FreetextAnswer freetextAnswer = new FreetextAnswer(question, true);
+        }
+    }
+}
