+++ src/main/java/de/imi/mopat/model/enumeration/QuestionType.java
@@ -1,5 +1,7 @@
 package de.imi.mopat.model.enumeration;
 
+import de.imi.mopat.controller.strategy.*;
+
 import java.util.HashMap;
 import java.util.Map;
 
@@ -8,10 +10,18 @@ import java.util.Map;
  */
 public enum QuestionType {
 
-    MULTIPLE_CHOICE("MULTIPLE_CHOICE"), SLIDER("SLIDER"), NUMBER_CHECKBOX(
-        "NUMBER_CHECKBOX"), NUMBER_CHECKBOX_TEXT("NUMBER_CHECKBOX_TEXT"), DROP_DOWN(
-        "DROP_DOWN"), FREE_TEXT("FREE_TEXT"), INFO_TEXT("INFO_TEXT"), NUMBER_INPUT(
-        "NUMBER_INPUT"), DATE("DATE"), IMAGE("IMAGE"), BODY_PART("BODY_PART"), BARCODE("BARCODE");
+    MULTIPLE_CHOICE("MULTIPLE_CHOICE", new MultipleChoiceOrDropdownStrategy()),
+    SLIDER("SLIDER", new SliderOrNumCheckBoxStrategy()),
+    NUMBER_CHECKBOX("NUMBER_CHECKBOX", new SliderOrNumCheckBoxStrategy()),
+    NUMBER_CHECKBOX_TEXT("NUMBER_CHECKBOX_TEXT", new NumberCheckBoxTextStrategy()),
+    DROP_DOWN("DROP_DOWN", new MultipleChoiceOrDropdownStrategy()),
+    FREE_TEXT("FREE_TEXT", new FreeTextOrBarcodeStrat()),
+    INFO_TEXT("INFO_TEXT", new DoNothing()),
+    NUMBER_INPUT("NUMBER_INPUT", new NumberInputStrat()),
+    DATE("DATE", new DateStrat()),
+    IMAGE("IMAGE", new ImageStrat()),
+    BODY_PART("BODY_PART", new BodyPartStrategy()),
+    BARCODE("BARCODE", new FreeTextOrBarcodeStrat());
     private final String textValue;
     private static final Map<String, QuestionType> stringToEnum = new HashMap<String, QuestionType>();
 
@@ -22,8 +32,15 @@ public enum QuestionType {
         }
     }
 
-    QuestionType(final String textValue) {
+    private final CreateOrUpdateAnswerStrategy strategy;
+
+    QuestionType(final String textValue, final CreateOrUpdateAnswerStrategy strategy) {
         this.textValue = textValue;
+        this.strategy = strategy;
+    }
+
+    public CreateOrUpdateAnswerStrategy getStrategy() {
+        return this.strategy;
     }
 
     @Override
